backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-docx==1.2.0
PyPDF2==3.0.1
pillow==10.2.0
requests==2.32.0
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.4
